+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.1.2 / ...
 
-* ...
+* Fix Docker tags on release (was not updating the docker.io tags)
 
 ## v0.1.1 / 2017-10-29